JMLXAIProject - LIME and SHAP for Julia

This package implements the explainable AI methods LIME and SHAP using XAIBase.jl. JMLXAIproject provides explanations for image and text input and is visualized as a heatmap.

To use the package please add the following code to your environment

using Pkg
Pkg.add(url="https://github.com/e-strauss/JML_XAI_Project.jl")

Documentation for JMLXAIProject.

JML_XAI_Project.data_labels

JML_XAI_Project.default_segmentation_function

JML_XAI_Project.euclidian_distance

JML_XAI_Project.explain_instance

JML_XAI_Project.data_labels — Function

Generates perturbed versions of a given image by turning superpixels on or off,using a specified segmentation map. It then predicts the class probabilities for these perturbed images using a provided classifier function. The function returns a tuple containing the binary matrix of perturbed images (data) and their corresponding prediction probabilities (labels). This is useful for techniques like LIME to understand and explain model predictions locally.

JML_XAI_Project.default_segmentation_function — Method

return image segmantation function, if no function was passed originally based on Scikit-Image implementation julia adaptations:

quickshift

package: ??? (docs: ???)

explaination: ???

python packages can be used in julia, it's therefore possible to use the scikit-image library if desired

slic

code: https://github.com/Cuda-Chen/SLIC.jl/tree/master (docs: NOT EVEN AN INOFFICIAL PACKAGE)

code seems to work, but spelling mistakes in the original and takes forever

explaination: https://cuda-chen.github.io/image%20processing/2020/07/29/slic-in-julia.html

felzenszwalb

package: ImageSegmentation (docs: https://juliaimages.org/v0.21/imagesegmentation/)

explaination: https://www.analyticsvidhya.com/blog/2021/05/image-segmentation-with-felzenszwalbs-algorithm/

comparision: https://scikit-image.org/docs/stable/autoexamples/segmentation/plotsegmentations.html

Args: algotype: string, segmentation algorithm among the following: 'quickshift', 'slic', 'felzenszwalb' targetparams: dict, algorithm parameters (valid model paramters as define in Scikit-Image documentation)

JML_XAI_Project.euclidian_distance — Method

calculates the euclidian distance between each column vector in input matrix A and the column vectors in input matrix B

Args: A: matrix (m,n) B: matrix (m,n) or (1,n) Returns: distance: 1-d array of distances

JML_XAI_Project.explain_instance — Function

Generates explanations for a prediction.

First, we generate neighborhood data by randomly perturbing features from the instance (see __datainverse). We then learn locally weighted linear models on this neighborhood data to explain each of the classes in an interpretable way (see limebase.py).

Args: image: 3 dimension RGB image. If this is only two dimensional, we will assume it's a grayscale image and call gray2rgb. classifierfn: classifier prediction probability function, which takes a numpy array and outputs prediction probabilities. For ScikitClassifiers , this is classifier.predictproba. labels: iterable with labels to be explained. hidecolor: If not None, will hide superpixels with this color. Otherwise, use the mean pixel color of the image. toplabels: if not None, ignore labels and produce explanations for the K labels with highest prediction probabilities, where K is this parameter. numfeatures: maximum number of features present in explanation numsamples: size of the neighborhood to learn the linear model batchsize: batch size for model predictions distancemetric: the distance metric to use for weights. modelregressor: sklearn regressor to use in explanation. Defaults to Ridge regression in LimeBase. Must have modelregressor.coef_ and 'sampleweight' as a parameter to modelregressor.fit() segmentationfn: SegmentationAlgorithm, wrapped skimage segmentation function randomseed: integer used as random seed for the segmentation algorithm. If None, a random integer, between 0 and 1000, will be generated using the internal random number generator.

Returns: An ImageExplanation object (see lime_image.py) with the corresponding explanations.