Base Functions

JML_XAI_Project.LIME — Type

struct LIME{M} <: AbstractXAIMethod

The LIME (Local Interpretable Model-agnostic Explanations) struct is used to create an instance of the LIME method for explainable AI (XAI). This method provides local explanations for the predictions of any machine learning model by approximating the model's behavior in the vicinity of a specific input.

Fields

model::M: The machine learning model to be explained. The model should be callable with an input to produce an output.

JML_XAI_Project.data_labels — Function

Generates perturbed versions of a given image by turning superpixels on or off,using a specified segmentation map. It then predicts the class probabilities for these perturbed images using a provided classifier function. The function returns a tuple containing the binary matrix of perturbed images (data) and their corresponding prediction probabilities (labels). This is useful for techniques like LIME to understand and explain model predictions locally.

JML_XAI_Project.default_segmentation_function — Method

return image segmantation function, if no function was passed originally based on Scikit-Image implementation julia adaptations:

quickshift

package: ??? (docs: ???)

explaination: ???

python packages can be used in julia, it's therefore possible to use the scikit-image library if desired

slic

code: https://github.com/Cuda-Chen/SLIC.jl/tree/master (docs: NOT EVEN AN INOFFICIAL PACKAGE)

code seems to work, but spelling mistakes in the original and takes forever

explaination: https://cuda-chen.github.io/image%20processing/2020/07/29/slic-in-julia.html

felzenszwalb

package: ImageSegmentation (docs: https://juliaimages.org/v0.21/imagesegmentation/)

explaination: https://www.analyticsvidhya.com/blog/2021/05/image-segmentation-with-felzenszwalbs-algorithm/

comparision: https://scikit-image.org/docs/stable/autoexamples/segmentation/plotsegmentations.html

Args: algotype: string, segmentation algorithm among the following: 'quickshift', 'slic', 'felzenszwalb' targetparams: dict, algorithm parameters (valid model paramters as define in Scikit-Image documentation)

JML_XAI_Project.euclidian_distance — Method

calculates the euclidian distance between each column vector in input matrix A and the column vectors in input matrix B

Args: A: matrix (m,n) B: matrix (m,n) or (1,n) Returns: distance: 1-d array of distances

JML_XAI_Project.explain_instance — Function

Generates explanations for a prediction.

First, we generate neighborhood data by randomly perturbing features from the instance (see __datainverse). We then learn locally weighted linear models on this neighborhood data to explain each of the classes in an interpretable way (see limebase.py).

Args: image: 3 dimension RGB image. If this is only two dimensional, we will assume it's a grayscale image and call gray2rgb. classifierfn: classifier prediction probability function, which takes a numpy array and outputs prediction probabilities. For ScikitClassifiers , this is classifier.predictproba. labels: iterable with labels to be explained. hidecolor: If not None, will hide superpixels with this color. Otherwise, use the mean pixel color of the image. toplabels: if not None, ignore labels and produce explanations for the K labels with highest prediction probabilities, where K is this parameter. numfeatures: maximum number of features present in explanation numsamples: size of the neighborhood to learn the linear model batchsize: batch size for model predictions distancemetric: the distance metric to use for weights. modelregressor: sklearn regressor to use in explanation. Defaults to Ridge regression in LimeBase. Must have modelregressor.coef_ and 'sampleweight' as a parameter to modelregressor.fit() segmentationfn: SegmentationAlgorithm, wrapped skimage segmentation function randomseed: integer used as random seed for the segmentation algorithm. If None, a random integer, between 0 and 1000, will be generated using the internal random number generator.

Returns: An ImageExplanation object (see lime_image.py) with the corresponding explanations.

JML_XAI_Project.explain_instance_with_data — Function

Takes perturbed data, labels and distances, returns explanation.

Parameters

neighborhood_data: perturbed data

neighborhood_labels: corresponding perturbed labels. should have as many columns as the number of possible labels.

distances: distances to original data point.

kernel_fn: (similiarity) kernel function that transforms an array of distances into an array of proximity values (floats)

label: label for which we want an explanation

num_features: maximum number of features in explanation

model_regressor: sklearn regressor to use in explanation. Defaults to Ridge regression if None. Must have modelregressor.coef and 'sampleweight' as a parameter to modelregressor.fit()

JML_XAI_Project.feature_selection — Method

feature_selection(X::Matrix, y::Vector, max_feat::Int) -> ReturnType

Selects features for the model using LARS with Lasso [https://tibshirani.su.domains/ftp/lars.pdf] s.t. len(selectedfeatures) <= maxfeat Use LARS package: https://github.com/simonster/LARS.jl

Python reference: nonzero = range(weighteddata.shape[1]) coefs = generatelarspath(X, y) for i in range(len(coefs.T) - 1, 0, -1): nonzero = coefs.T[i].nonzero()[0] if len(nonzero) <= numfeatures: return nonzero

Parameters

X: weighted feature

y: weighted labels

Returns

indices of selected features

JML_XAI_Project.sample_data — Method

sample_data(x0::Vector, x0_pertubed::Vector, H, D, model)

Returns the neighborhood_data (perturbed data, first element is the original data point) by sampling around x0 and evaluate the model to generate the corresponding perturbed labels.

Citation from Paper [https://arxiv.org/pdf/1602.04938]: We sample instances around x0pertubed by drawing nonzero elements of x0pertubed uniformly at random (where the number of such draws is also uniformly sampled). We sample instances both in the vicinity of x0 and far away from x (measured by the distance metric).

Parameters

x0: original input

x0_pertubed: interpretable representation of original input

H: interpretable representation mapping function, s.t. h(x0_pertubed) = x0

D: distance metric which calculates the distance of two points in the origial, non-interpretable space

model: model which works on the non-interpretable original input

Returns

neighborhood_data: sampled perturbed data (first element is x0_pertubed)

neighborhood_labels: corresponding perturbed calculated by the model and mapping function H

distances: distances of each sampled point to original data point calculated using the distance metric D and mapping function H

JML_XAI_Project.train_ridge_regressor — Method

trainridgeregressor(X::Matrix, y::Vector, weights::Vector)

Returns the trained simplified linear model using ridge regression: model = Ridge(alpha=1, fitintercept=True, randomstate=self.randomstate) model.fit(X,y, sampleweight=weights) return model

JML_XAI_Project.weighted_data — Method

normalize_data(X::Matrix, y::Vector, weights::Vector)

Returns the weight normalization of X (data) and y (label) using the weight vector y. Xnorm = ((X - np.average(X, axis=0, weights=weights)) * np.sqrt(weights[:, np.newaxis])) Ynorm = ((y - np.average(y, weights=weights)) * np.sqrt(weights))