Base Functions

XAIBase Wrapper for calling LIME

JML_XAI_Project.LIME — Type

struct LIME{M} <: AbstractXAIMethod

The LIME (Local Interpretable Model-agnostic Explanations) struct is used to create an instance of the LIME method for explainable AI (XAI). This method provides local explanations for the predictions of any machine learning model by approximating the model's behavior in the vicinity of a specific input.

Fields

model::M: The machine learning model to be explained. The model should be callable with an input to produce an output.

JML_XAI_Project.LIME — Method

(method::LIME)(input, output_selector::AbstractOutputSelector)

This function applies the LIME (Local Interpretable Model-agnostic Explanations) method to a given input and returns an explanation of the model's prediction.

Arguments

input: The input data for the model, typically a 4D array (e.g., image data with dimensions height x width x channels x batch size).

output_selector::AbstractOutputSelector: An object that selects specific outputs from the model's predictions.

Returns

Explanation: An object containing the following fields:

val: The explanation value generated by the explain_instance function.

output: The output of the model for the given input.

output_selection[1]: The selected output used for generating the explanation.

:MyMethod: The method used for generating the explanation.

:attribution: The type of explanation provided (attribution).

extras: Additional information related to the explanation (currently set to nothing).

Generation of a new interpretable data set

Conversion of the input image into mulitple disturbed interpretable representation

JML_XAI_Project.explain_instance — Function

explain_instance(image, classifier_fn, output_selection, num_features=16, num_samples=64, batch_size=5, distance_metric="cosine",)

Generates explanations for a prediction.

we generate neighborhood data by randomly perturbing features from the instance.

We then learn locally weighted linear models on this neighborhood data to explain each of the classes in an interpretable way.

Parameters

image: 4 dimension RGB image.

classifier_fn: classifier prediction probability function, which takes a image Matrix and outputs prediction probabilities.

labels: iterable with labels to be explained.

num_features: maximum number of features present in explanation, default = 16.

num_samples: size of the neighborhood (perturbed images) to learn the linear model, default = 64.

batch_size: batch size for model predictions, default = 5.

distance_metric: the distance metric to use for weights, default = 'cosine'

Returns:

An ImageExplanation object with the corresponding explanations.

JML_XAI_Project.default_segmentation_function — Function

function default_segmentation_function(algo_type::String)

return image segmantation function, if no function was passed originally based on Scikit-Image implementation julia adaptations:

quickshift

package: ??? (docs: ???)

explaination: ???

python packages can be used in julia, it's therefore possible to use the scikit-image library if desired

slic

code: https://github.com/Cuda-Chen/SLIC.jl/tree/master (docs: NOT EVEN AN INOFFICIAL PACKAGE)

code seems to work, but spelling mistakes in the original and takes forever

explaination: https://cuda-chen.github.io/image%20processing/2020/07/29/slic-in-julia.html

felzenszwalb

package: ImageSegmentation (docs: https://juliaimages.org/v0.21/imagesegmentation/)

explaination: https://www.analyticsvidhya.com/blog/2021/05/image-segmentation-with-felzenszwalbs-algorithm/

comparision: https://scikit-image.org/docs/stable/autoexamples/segmentation/plotsegmentations.html

Parameters

algo_type: string, segmentation algorithm among the following: 'quickshift', 'slic', 'felzenszwalb'

target_params: dict, algorithm parameters (valid model paramters as define in Scikit-Image documentation)

Returns

JML_XAI_Project.cosine_similiarity — Function

cosine_similarity(A::AbstractArray, B::AbstractArray)

Computes the cosine similarity between corresponding rows of two arrays.

Parameters

A::AbstractArray: The first input array. Each row represents a vector.

B::AbstractArray: The second input array. Each row represents a vector. Must have the same number of columns as A.

Returns

AbstractArray: An array of cosine similarities between corresponding rows of A and B.

JML_XAI_Project.pairwise_distance — Function

pairwise_distance(A::AbstractArray, B::AbstractArray; method="cosine")

Computes the pairwise distance between corresponding rows of two arrays using the specified distance metric.

Parameters

A::AbstractArray: The first input array. Each row represents a vector.

B::AbstractArray: The second input array. Each row represents a vector. Must have the same number of columns as A.

method::String="cosine": The distance metric to use. Options are "cosine" for cosine similarity or "euclidean" for Euclidean distance. Defaults to "cosine".

Returns

AbstractArray: An array of distances between corresponding rows of A and B.

JML_XAI_Project.data_labels — Function

function data_labels(image, fudged_image, segments, classifier_fn, num_samples, batch_size=10)

Generates perturbed versions of a given image by turning superpixels on or off,using a specified segmentation map. It then predicts the class probabilities for these perturbed images using a provided classifier function. The function returns a tuple containing the binary matrix of perturbed images (data) and their corresponding prediction probabilities (labels). This is useful for techniques like LIME to understand and explain model predictions locally.

Parameters

image::Matrix{RGB{Float32}}: The original input image.

fudged_image::Matrix{RGB{Float32}}: The fudged image with mean colors for segments.

segments::Matrix{Int}: The segmentation map where each segment is labeled with an integer.

classifier_fn::Function: A function that takes a batch of images and returns their predicted labels.

num_samples::Int: The number of samples to generate.

batch_size::Int=10: The size of the batches to process through the classifier function. Defaults to 10.

Returns

data::Matrix{Int}: A binary matrix where each row indicates which features (segments) are active.

labels::Matrix{Float64}: A matrix of classifier predictions for the corresponding images in data.

JML_XAI_Project.euclidian_distance — Function

function euclidian_distance(A,B)

calculates the euclidian distance between each column vector in input matrix A and the column vectors in input matrix B

Parameters:

A: matrix (m,n)

B: matrix (m,n) or (1,n)

Returns:

distance: 1-d array of distances

JML_XAI_Project.exponential_kernel — Function

exponential_kernel(d::AbstractArray; kernel_width=0.25)

Computes the exponential kernel for a given array of distances.

Parameters

d::AbstractArray: An array of distances.

kernel_width::Float64=0.25: The width of the kernel. Defaults to 0.25.

Returns

AbstractArray: An array of kernel values computed from the input distances.

JML_XAI_Project.create_fudged_image — Function

create_fudged_image(img::Matrix{RGB{Float32}}, seg_map::Matrix{Int})

Creates a "fudged" image where the pixels of each segment are replaced by the mean color of that segment.

Parameters

img: The input image as a matrix of RGB colors with floating point values.

seg_map: The segmentation map containing the segment labels for each pixel.

Returns

Matrix{RGB{Float32}}: A new image where the pixels of each segment are replaced by the mean color of that segment.

Feature selection and train simplified model on the interpretable data set

JML_XAI_Project.explain_instance_with_data — Function

function explain_instance_with_data(neighborhood_data, neighborhood_labels, distances, label, num_features, kernel_fn = (x) -> 1 .- x)

Takes perturbed data, labels and distances, returns explanation.

Parameters

neighborhood_data: perturbed data

neighborhood_labels: corresponding perturbed labels. should have as many columns as the number of possible labels.

distances: distances to original data point.

kernel_fn: (similiarity) kernel function that transforms an array of distances into an array of proximity values (floats)

label: label for which we want an explanation

num_features: maximum number of features in explanation

model_regressor: sklearn regressor to use in explanation. Defaults to Ridge regression if None. Must have modelregressor.coef and 'sampleweight' as a parameter to modelregressor.fit()

Returns

JML_XAI_Project.weighted_data — Function

normalize_data(X::Matrix, y::Vector, weights::Vector)

Returns the weight normalization of features and labels using a weight vector.

Parameters

X: features

y: labels

-weights: weight vector

Returns

JML_XAI_Project.train_ridge_regressor — Function

train_ridge_regressor(X::Matrix, y::Vector, weights::Vector)

Returns the trained simplified linear model using ridge regression:

Parameters

X:

y:

Returns

simplified linear model

JML_XAI_Project.feature_selection — Function

feature_selection(X::Matrix, y::Vector, max_feat::Int) -> ReturnType

Selects features for the model using LARS with Lasso [https://tibshirani.su.domains/ftp/lars.pdf] s.t. len(selectedfeatures) <= maxfeat Use LARS package: https://github.com/simonster/LARS.jl

Parameters

X: weighted feature

y: weighted labels

Returns

indices of selected features

Index

JML_XAI_Project.LIME

JML_XAI_Project.LIME

JML_XAI_Project.cosine_similiarity

JML_XAI_Project.create_fudged_image

JML_XAI_Project.data_labels

JML_XAI_Project.default_segmentation_function

JML_XAI_Project.euclidian_distance

JML_XAI_Project.explain_instance

JML_XAI_Project.explain_instance_with_data

JML_XAI_Project.exponential_kernel

JML_XAI_Project.feature_selection

JML_XAI_Project.pairwise_distance

JML_XAI_Project.train_ridge_regressor

JML_XAI_Project.weighted_data